% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{asp}
\alias{asp}
\title{asp}
\usage{
asp(x)
}
\arguments{
\item{x}{meme object}
}
\value{
asp ratio
}
\description{
aspect ratio of meme
}
\examples{
f <- system.file("icon.png", package="meme")
x <- meme(f, "meme", "produced by meme package")
asp(x)
}
\author{
guangchuang yu
}
