% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_import.R
\name{font_pokemon}
\alias{font_pokemon}
\alias{font_import}
\title{font_pokemon}
\usage{
font_pokemon()

font_import(path = ".")
}
\arguments{
\item{path}{folder of fonts}
}
\description{
load fonts
}
\author{
Guangchuang Yu
}
