% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grid.echo.meme}
\alias{grid.echo.meme}
\title{grid.echo.meme}
\usage{
\method{grid.echo}{meme}(x = NULL, newpage = TRUE, prefix = NULL, ...)
}
\arguments{
\item{x}{meme object}

\item{newpage}{not use in this method}

\item{prefix}{not use in this method}

\item{...}{not use in this method}
}
\value{
meme object
}
\description{
grid.echo method for meme object
}
\details{
using the solution <https://stackoverflow.com/a/18551959/3844636>
to not importing grid.echo generics from gridGraphics for easy installation,
since gridGraphics has system requirement of ImageMagick <https://github.com/cran/gridGraphics/blob/master/DESCRIPTION#L14>.

this method is just a hack for making cowplot to support meme object, and not used elsewhere.
}
\author{
Guangchuang Yu
}
