% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{+.meme}
\alias{+.meme}
\title{Setting meme parameter}
\usage{
\method{+}{meme}(e1, e2)
}
\arguments{
\item{e1}{meme object}

\item{e2}{aes()}
}
\description{
Setting meme parameter
}
\examples{
f <- system.file("angry8.jpg", package="meme")
meme(f, "code", "all the things!") + aes(color="firebrick")
}
