\name{meme-package}
\alias{meme-package}
\docType{package}
\title{Meme Generation in R}
\description{Generate memes on a variety of meme generators from R}
\details{Use \href{https://imgflip.com/}{imgflip}, \href{http://memegenerator.net/}{memegenerator}, or \href{http://memecaptain.com/}{memecaptain} to generate memes from a template image. Use of the imgflip and memegenerator generators requires account registration.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}

