% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{meme}
\alias{meme}
\title{meme}
\usage{
meme(img, upper = "", lower = "", size = "auto", color = "white",
  font = "Impact", vjust = 0.1, bgcolor = "black", r = 0.01)
}
\arguments{
\item{img}{path or url}

\item{upper}{upper text}

\item{lower}{lower text}

\item{size}{size of text}

\item{color}{color of text}

\item{font}{font family of text}

\item{vjust}{vertical adjustment of captions}

\item{bgcolor}{background color of shadow text}

\item{r}{ratio of shadow text}
}
\value{
grob object
}
\description{
create meme
}
\examples{
f <- system.file("angry8.jpg", package="meme")
meme(f, "code", "all the things!", font = "Helvetica")
}
\author{
guangchuang yu
}
