% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{meme_save}
\alias{meme_save}
\title{meme_save}
\usage{
meme_save(x, file, width = NULL, height = NULL, ...)
}
\arguments{
\item{x}{meme output}

\item{file}{output file}

\item{width}{width of graph}

\item{height}{height of graph}

\item{...}{additional arguments for ggsave}
}
\description{
save meme plot
}
\examples{
f <- system.file("icon.png", package="meme")
x <- meme(f, "code", "all the things!")
outfile <- tempfile(fileext = ".png")
meme_save(x, outfile)
}
\author{
guangchuang yu
}
