% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob.R
\name{memeGrob}
\alias{memeGrob}
\title{memeGrob}
\usage{
memeGrob(x)
}
\arguments{
\item{x}{meme object}
}
\value{
grob object
}
\description{
convert meme to grob object
}
\examples{
f <- system.file("angry8.jpg", package="meme")
x <- meme(f, "code", "all the things!")
memeGrob(x)
}
\author{
guangchuang yu
}
