% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{mm_caption}
\alias{mm_caption}
\title{mm_caption}
\usage{
mm_caption(upper = NULL, lower = NULL, ...)
}
\arguments{
\item{upper}{upper caption}

\item{lower}{lower caption}

\item{...}{additional parameters to set caption}
}
\value{
meme object
}
\description{
add caption layer for meme
}
\author{
guangchuang yu
}
