% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{plot.meme}
\alias{plot.meme}
\title{plot method for meme object}
\usage{
\method{plot}{meme}(x, size = NULL, color = NULL, font = NULL,
  upper = NULL, lower = NULL, dev.new = TRUE, ...)
}
\arguments{
\item{x}{meme object}

\item{size}{size of text}

\item{color}{color of text}

\item{font}{font family of text}

\item{upper}{upper text}

\item{lower}{lower text}

\item{dev.new}{wheter open new device}

\item{...}{other arguments not used by this method}
}
\description{
plot method for meme object
}
