% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.meme}
\alias{print.meme}
\alias{plot.meme}
\title{print method for meme object}
\usage{
\method{print}{meme}(x, size = NULL, color = NULL, font = NULL,
  upper = NULL, lower = NULL, vjust = NULL, newpage = is.null(vp),
  vp = NULL, ...)

\method{plot}{meme}(x, size = NULL, color = NULL, font = NULL,
  upper = NULL, lower = NULL, vjust = NULL, newpage = is.null(vp),
  vp = NULL, ...)
}
\arguments{
\item{x}{meme object}

\item{size}{size of text}

\item{color}{color of text}

\item{font}{font family of text}

\item{upper}{upper text}

\item{lower}{lower text}

\item{vjust}{vertical adjustment ratio}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\description{
print method for meme object
}
\examples{
f <- system.file("angry8.jpg", package="meme")
x <- meme(f, "code", "all the things!")
print(x)
}
