% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\docType{data}
\name{+.meme}
\alias{+.meme}
\alias{\%+\%}
\title{Setting meme parameter}
\format{An object of class \code{character} of length 1.}
\usage{
\method{+}{meme}(e1, e2)

%+%
}
\arguments{
\item{e1}{meme object}

\item{e2}{aes()}
}
\description{
Setting meme parameter
}
\keyword{datasets}
