% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{meme}
\alias{meme}
\title{meme}
\usage{
meme(img, upper = "", lower = "", size = "auto", color = "white",
  font = "Helvetica", plot = TRUE)
}
\arguments{
\item{img}{path or url}

\item{upper}{upper text}

\item{lower}{lower text}

\item{size}{size of text}

\item{color}{color of text}

\item{font}{font family of text}

\item{plot}{boolean, one of TRUE or FALSE}
}
\value{
grob object
}
\description{
create meme
}
\examples{
\dontrun{
u <- "http://www.happyfamilyneeds.com/wp-content/uploads/2017/08/angry8.jpg"
meme(u, "code", "all the things!")
}
}
\author{
guangchuang yu
}
